requirements.txt
transformers
huggingface_hub
nltk
gradio
pyvis
evaluate
python-dotenv
# git+https://github.com/huggingface/peft.git
trl
bitsandbytes
beautifulsoup4
scrapy
pandas
numpy
torch
gradio
matplotlib
seaborn
plotly
spacy
datasets



